ml_service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1

# Data & ML
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
matplotlib==3.9.0
seaborn==0.13.2

# PyTorch — CPU build; swap for the CUDA variant if you have a GPU
torch==2.3.1
torchvision==0.18.1

# Utilities
joblib==1.4.2
